+++ docs/CHANGELOG.md
@@ -5,7 +5,7 @@
 
 - Fixed bug with app not exiting on last window.
 - Remove trial text since we can't currently offer trials
-- 
+- WAY better error message when we can't connect to Anki.
 
 # ....